columns typed as DTYPE_DATE now return datetime.date objects
