+++ python/perspective/perspective/tests/table/test_table_numpy.py
@@ -15,13 +15,6 @@ from perspective.table import Table
 from datetime import date, datetime
 
 
-def _to_timestamp(obj):
-    if six.PY2:
-        return int((time.mktime(obj.timetuple()) + obj.microsecond / 1000000.0))
-    else:
-        return obj.timestamp()
-
-
 class TestTableNumpy(object):
     def test_empty_table(self):
         tbl = Table([])
@@ -194,8 +187,8 @@ class TestTableNumpy(object):
             "b": date
         }
         assert tbl.view().to_dict() == {
-            "a": [datetime(2019, 7, 11, 0, 0)],
-            "b": [datetime(2019, 7, 12, 0, 0)]
+            "a": [date(2019, 7, 11)],
+            "b": [date(2019, 7, 12)]
         }
 
     def test_table_np_datetime(self):
@@ -618,7 +611,7 @@ class TestTableNumpy(object):
             "a": date
         })
         table.update(df)
-        assert table.view().to_dict()["a"] == [datetime(2019, 8, 15, 0, 0), None, datetime(2019, 8, 16, 0, 0)]
+        assert table.view().to_dict()["a"] == [date(2019, 8, 15), None, date(2019, 8, 16)]
 
     def test_table_numpy_from_schema_datetime(self):
         data = [datetime(2019, 7, 11, 12, 30, 5), None, datetime(2019, 7, 11, 13, 30, 5), None]
@@ -629,8 +622,8 @@ class TestTableNumpy(object):
         table.update(df)
         assert table.view().to_dict()["a"] == data
 
-    def test_table_numpy_from_schema_datetime_timestamp_s(self):
-        data = [_to_timestamp(datetime(2019, 7, 11, 12, 30, 5)), np.nan, _to_timestamp(datetime(2019, 7, 11, 13, 30, 5)), np.nan]
+    def test_table_numpy_from_schema_datetime_timestamp_s(self, util):
+        data = [util.to_timestamp(datetime(2019, 7, 11, 12, 30, 5)), np.nan, util.to_timestamp(datetime(2019, 7, 11, 13, 30, 5)), np.nan]
         df = {"a": np.array(data)}
         table = Table({
             "a": datetime
@@ -643,11 +636,11 @@ class TestTableNumpy(object):
             None
         ]
 
-    def test_table_numpy_from_schema_datetime_timestamp_ms(self):
+    def test_table_numpy_from_schema_datetime_timestamp_ms(self, util):
         data = [
-            _to_timestamp(datetime(2019, 7, 11, 12, 30, 5)) * 1000,
+            util.to_timestamp(datetime(2019, 7, 11, 12, 30, 5)) * 1000,
             np.nan,
-            _to_timestamp(datetime(2019, 7, 11, 13, 30, 5)) * 1000,
+            util.to_timestamp(datetime(2019, 7, 11, 13, 30, 5)) * 1000,
             np.nan
         ]
